extract product relationships to a new trait
